requirements.txt
Flask
pymongo
beautifulsoup4
python-dotenv